requirements.txt
flask
flask_restful
jinja2
pyyaml
uwsgi
flask_cors
jinja2_ansible_filters
flask_swagger
pytest
netaddr
